Add files via upload
Building the Mining Claims page layout and design, gonna be a bit still but wanted to push it to github.  needs a button back and forth to home page or home page needs one back to it.  Has the iframe now with google maps and is working. everything seems to be fitting in it's spots now so can adjust and design more. Added a locations folder where each claim will be seperated by it's own folder, which will be directed to it's own individual page, or so we hope lol