Add files via upload
Trying to fix this m&tamask wallet, i think it should connect to wallets now. it was trying to connect to a class and not id, plus had two class names for styling, but it didn't like that on the javascript end.  works now on my end tho on this side