Add files via upload
Fixed 2nd wallet connect half way down index.html the homepage.  The styling needs to be adjusted.  Added an extra div with an id of started-items-item, the same as the class, but it was easier to just wrap an empty div around and give it an id instead of changing all the other class names or i dont know just seemed to work for the moment.