Add files via upload
Tried to move claim_01 and move to diffrent folder, i think it actually deleted it lol, but i had a back up. put new file claim_01.html into a new folder called Mine which has the html for each claim and the style will be in the css folder, was trying to keep everything in subfolders of maybe sub folders but again it had room for error.  more tests coming through