Add files via upload
Test 02 removed locations folder which was original folder that had the claim_01 page, deleted that cause there was another subfolder, than moved and remade it again, but it wasnt deleted correctly i guess or something, maybe cause i made the new file in the same name. subfolder try not to do them.  Added package.jsn file to see if that heps with error directory