Add files via upload
Site is Done,just dialing in mobile view a bit.  the yellow notifications seem to have something to do with m&tamask  is what ol google says, but will research in the future.  Hope y'all enjoy it, ya just need to fork the repository with the name Juce on it and Viola.  There's some tutorials in the README on how to do it or can walk ya through.  thank ya