Add files via upload
Added spot for Logo next to name at Top. Renamed claim_01.css to claims.css so it can be template for other pages in that category. Added claim_2.html through claim_05.html as some more Extra Files & so they have examples + extra content for them, they approved design today on phone and said was good it will work. just to have records in here for it as well:) We will be forking this repo today and moving the site to them. Cheers!.  just Added contract address spot with a copy button so anyone can just copy address easily with click of a button