Add files via upload
Adding to anchor tag to images to claims on index page so they can go back and forth to those claims as well.  made 5 starter claims off the template to test it out a few commits ago but now they are linked in more places.  Organizing links on pages so they again can move around from page to page more easily. Added a hidden game for the adventurous ones! Nice fun feature for the client kinda like the cherry on top ya know.